# api/swagger/gardenbot-api.yaml
swagger: "2.0"

info:
  title: "Gardenbot API"
  version: "1.0"

basePath: /v1.0

paths:
  /watering/{seconds}:
    post:
      summary: Waters the plants
      description: Runs the pump for the given time (in seconds)
      operationId: classes.gardenbot-api.water_plants
      produces:
        - text/plain;
      responses:
        200:
          description: watering response
          schema:
            type: string
          examples:
            "text/plain": "Watering the plants..."
      parameters:
        - name: seconds
          in: path
          description: Time in seconds, the plants shall be watered
          required: true
          type: string
        - name: API-Key
          in: header
          type: string
          required: true
  /check:
    post:
      summary: Checks if soil is wet enough or not
      description: Checks wether the soil is wet or not
      operationId: classes.gardenbot-api.check_moisture
      produces:
        - text/plain;
      responses:
        200:
          description: watering response
          schema:
            type: string
          examples:
            "text/plain": "Watering the plants..."
      parameters:
        - name: API-Key
          in: header
          type: string
          required: true
  /water_history:
    get:
      summary: Returns watering activities
      description: Returns the watering activities of the last 5 days
      operationId: classes.gardenbot-api.get_water_history
      produces:
      - application/json;
      responses:
        200:
          description: history response
          schema:
            type: string
          examples:
            "text/json": "[
                              [
                                '2018-04-11', 1
                              ]
                          ],
                          [
                              [
                                '2018-04-12', 3
                              ]
                          ]"
      parameters:
        - name: API-Key
          in: header
          type: string
          required: true
#  /sensordata_history:
#    get:
#      summary: Returns sensordata of the past
#      description: Returns timestamp, temperature, moisture, fertility, light and battery of the past
#      operationId: classes.gardenbot-api.get_sensor_history
#      produces:
#      - application/json;
#      responses:
#        200:
#          description: history response
#          schema:
#            type: string
#          examples:
#            "text/plain": test
#      parameters:
#        - name: API-Key
#          in: header
#          type: string
#          required: true
  /sensordata:
    get:
      summary: Returns the current sensor data
      description: Returns current temperature, moisture, fertility, light and battery
      operationId: classes.gardenbot-api.get_sensor_data
      produces:
      - application/json;
      responses:
        200:
          description: positive
          schema:
            type: string
      parameters:
        - name: API-Key
          in: header
          type: string
          required: true
  /waterstatus:
    get:
      summary: Water indicator
      description: Returns true, if there's enough water in the bucket, otherwise false
      operationId: classes.gardenbot-api.get_water_status
      produces:
      - application/json;
      responses:
        200:
          description: water status response
          schema:
            type: string
          examples:
            "text/plain": "Showing water status"
      parameters:
        - name: API-Key
          in: header
          type: string
          required: true